Crate const_str

Compile-time string operations

MSRV: Rust 1.61.0

Macros

concat

Concatenates values into a string slice.

contains

Returns true if the given pattern matches a sub-slice of this string slice.

convert_ascii_case

Converts a string slice to a specified case. Non-ascii characters are not affected.

encode

Encode a string slice with a specified encoding.

encode_z

Encode a string slice with a specified encoding and append a nul character.

ends_with

Returns true if the given pattern matches a suffix of this string slice.

equal

Checks that two strings are equal.

from_utf8

Converts a byte string to a string slice

hex_bytes

Converts hexadecimal string slices to a byte array.

parse

Parse a value from a string slice.

repeat

Creates a new string slice by repeating a string slice n times.

replace

Replaces all matches of a pattern with another string slice.

split

Returns an array of substrings of a string slice, separated by characters matched by a pattern.

starts_with

Returns true if the given pattern matches a prefix of this string slice.

to_byte_array

Converts a string slice or a byte string to a byte array.

to_char_array

Converts a string slice into an array of its characters.

to_str

Converts a value to a string slice.