Crate const_str

Compile-time string operations

MSRV: Rust 1.64.0

Macros

compare

Compares two strings lexicographically.

concat

Concatenates values into a string slice.

concat_bytes

Concatenates values into a byte slice.

contains

Returns true if the given pattern matches a sub-slice of this string slice.

convert_ascii_case

Converts a string slice to a specified case. Non-ascii characters are not affected.

cstr

Converts a string slice to &CStr.

encode

Encode a string slice with a specified encoding.

encode_z

Encode a string slice with a specified encoding and append a nul character.

ends_with

Returns true if the given pattern matches a suffix of this string slice.

eq_ignore_ascii_case

Checks that two (string) slices are an ASCII case-insensitive match.

equal

Checks that two strings are equal.

from_utf8

Converts a byte string to a string slice

hex

Converts hexadecimal string slices to a byte array.

is_ascii

Checks if all characters in this (string) slice are within the ASCII range.

parse

Parse a value from a string slice.

raw_cstr

Converts a string slice to *const c_char.

repeat

Creates a new string slice by repeating a string slice n times.

replace

Replaces all matches of a pattern with another string slice.

sorted

Sorts string slices and returns a new array.

split

Returns an array of substrings of a string slice, separated by characters matched by a pattern.

squish

Splits the string by ASCII whitespaces, and then joins the parts with a single space.

starts_with

Returns true if the given pattern matches a prefix of this string slice.

strip_prefix

Returns a string slice with the prefix removed.

strip_suffix

Returns a string slice with the suffix removed.

to_byte_array

Converts a string slice or a byte string to a byte array.

to_char_array

Converts a string slice into an array of its characters.

to_str

Converts a value to a string slice.

unwrap

Unwraps a container, returns the content.